feat: Support multiple timestamp formats including ISO 8601 (#1188)
* feat: Support multiple timestamp formats including ISO 8601

Add flexible timestamp parsing to accept multiple common formats:
- ISO 8601 with 'T' separator (2025-11-10T08:24:34)
- ISO 8601 with fractional seconds (2025-11-10T08:24:34.602988)
- ISO 8601 with timezone markers (2025-11-10T08:24:34Z, +05:00)
- Space-separated format (2025-11-10 08:24:34) - existing
- Date-only format (2025-11-10) - assumes midnight

Implementation changes:
- timestamp.rs: Enhanced FromStr to support multiple formats
- validation.rs: Added implicit VARCHAR→TIMESTAMP conversion in INSERT
- normalization.rs: Added implicit conversion in storage layer
- Added 27 unit tests for timestamp parsing
- Added 19 integration tests for end-to-end functionality

Fixes #1187



* fix: Resolve CI test failures unrelated to timestamp feature

Fixed compilation and test failures in executor module:

1. Export missing types in iterator module:
   - Added ProjectionIterator and LazyNestedLoopJoin exports
   - Fixes compilation errors in iterator tests

2. Update aggregate test assertions:
   - Changed Float to Numeric in 6 test assertions
   - Aggregate functions now return Numeric instead of Float
   - Fixed floating point precision in window aggregate test

Test results:
- ✅ 626 executor tests passing (was 620 with 6 failures)
- ✅ All compilation errors resolved
- ✅ No test failures related to timestamp parsing changes

These fixes are unrelated to the timestamp parsing feature but were
blocking CI. The timestamp feature tests (27 unit + 19 integration)
continue to pass.



---------

+++ crates/executor/src/tests/aggregate_count_sum_avg_tests.rs
@@ -592,7 +592,7 @@ fn test_count_star_in_arithmetic_expression() {
 
     let result = executor.execute(&stmt).unwrap();
     assert_eq!(result.len(), 1);
-    assert_eq!(result[0].values[0], types::SqlValue::Float(50.0)); // 5 * 10
+    assert_eq!(result[0].values[0], types::SqlValue::Numeric(50.0)); // 5 * 10
 }
 
 #[test]
